Automatic qubit management works with gradients (#2262)
**Context:**
In automatic qubit management mode, a `quantum.num_qubits` operation is
inserted before terminal measurement ops to query their shape. This
breaks gradient transforms because these expect the cloned classical
parts of the circuit to be free of quantum ops, but the num qubits ops
are not removed there.

**Description of the Change:**
Add `MemoryRead` to `quantum.num_qubits` op: if it has no users, DCE
(and canonicalizer) can remove it.

**Benefits:**
Automatic qubit management works with gradients

[sc-105179]