Implement new `mlir_specs` Python function to get resource counts from MLIR passes (#2238)
**Context:**
There is currently no way to view the impact to runtime costs of a given
MLIR pass from the python frontend of PennyLane.

**Description of the Change:**
Creates a new `mlir_specs` function, in addition to a `specs_collect`
function which uses xDSL to allow inspection of compilation passes
written in MLIR.

**Benefits:**
Allows users to evaluate the impact on the IR of various MLIR passes.

**Possible Drawbacks:**
Not integrated with `qml.specs()`, this will be handled in a followup
PR.

**Related GitHub Issues:**
[sc-103510]
Migrated from https://github.com/PennyLaneAI/pennylane/pull/8660

See also the frontend integration for this PR:
https://github.com/PennyLaneAI/pennylane/pull/8606

---------

Co-authored-by: Mudit Pandey <mudit.pandey@xanadu.ai>
Co-authored-by: Mudit Pandey <18223836+mudit2812@users.noreply.github.com>
Co-authored-by: Mehrdad Malek <39844030+mehrdad2m@users.noreply.github.com>
Co-authored-by: David Ittah <dime10@users.noreply.github.com>