Fix for subroutine + for loop + qubit allocation (#2268)
**Context:**
Subroutine + for loop + qubit allocation fails.
This is because there's an integer check when judging whether a wire
label was dynamically allocated or not. This logic fails because
subroutines take in all wire labels as subroutine arguments, hence they
are tracers.

**Description of the Change:**
Relocate the check to only where it's needed (as a throwback to relay
truly `const` dynamically allocated wire indices, so subscope handlers
can remove them from their subjaxpr's `consts`).

**Benefits:**
Subroutine + for loop + qubit allocation works. This unblocks decomp
rules with dynamic allocation.

**Related GitHub Issues:** fixes #2263 

[sc-105198]