Add cli path fallback for rare wheel setups
The catalyst binary used to be shipped right within the package, under
catalyst/bin. This is no longer the case, and now pip will install the
binary into the dedicated binary folder for Python packages.
However, we may still encounter a configuration where the binary is
located in a bin folder next to the catalyst package folder,
specifically:
- after `make wheel` we'll have frontend/bin next to frontend/catalyst,
  although we rarely directly test this configuration since it's more
  of a setup to packaging the wheel.
- when trying to emulate the above configuration for a quick
  frontend-only editable install, e.g. by extracting wheel contents
  over a checked out catalyst repo.