[Gridsynth 2] Gridsynth decomposition algorithm implementation  (#2166)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new functions and code must be clearly commented and
documented.

- [ ] Ensure that code is properly formatted by running `make format`.
The latest version of black and `clang-format-20` are used in CI/CD to
check formatting.

- [ ] All new features must include a unit test.
Integration and frontend tests should be added to
[`frontend/test`](../frontend/test/),
Quantum dialect and MLIR tests should be added to
[`mlir/test`](../mlir/test/), and
Runtime tests should be added to [`runtime/tests`](../runtime/tests/).

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
This PR provides the implementation of the Ross-Selinger Gridsynth
algorithm to decompose RZ/PhaseShift gates into Clifford+T gates.

**Description of the Change:**
This PR primarily follows the Python implementation in
[PennyLane](https://github.com/PennyLaneAI/pennylane/tree/master/pennylane/ops/op_math/decompositions)
of the same algorithm. A new runtime directory
`runtime/lib/RSDecompRuntime` is added for the runtime component used
for the pass.

Note: This library requires the use of `Boost` library. This will be
automatically pulled in, unless it is found locally.

**Benefits:**
Gridsynth in Catalyst now supported. Note with the current
implementation, precision down to ~1e-7 is supported. Further than that,
precision issues will occur - a full multiprecision implementation will
be updated in the future.

**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-102151]