fix: add `use_both_frontends` fixture to all tests (#2305)
This feature needs to work with both frontends (i.e. with and without
capture). This unfortunately affected a few tests (tests were written
too specific to the legacy frontend).

⚠️ Order of operations in the IR (if operators are "parallel") are not
guaranteed to be the same with or without capture. That's why some tests
needed to be updated so they are agnostic to the order they appear.

[sc-105774]