switch to using cleanup (#2312)
**Context:**

`PlxprInterpreter` has a `cleanup` method to finalize everything and
cleanup variables as needed. This can be used to embed the decomposition
rules at the very end of decomposition graph interpretation.

**Description of the Change:**

Switch from solving the graph and embedding the rules on the first
measurement to doing so in the cleanup method before return.

**Benefits:**

Little clearer organization.

**Possible Drawbacks:**

**Related GitHub Issues:**

---------

Co-authored-by: Ali Asadi <10773383+maliasadi@users.noreply.github.com>